requirements.txt
requests==2.28.0
gspread==5.4.0
python-dotenv==0.20.0
sqlalchemy==1.4.37  # удалить
pandas==1.4.2
psycopg2